restructured events to util
