+++ theseus/src/lib.rs
@@ -20,4 +20,4 @@ pub use api::*;
 pub use error::*;
 pub use state::State;
 
-pub use launcher::events::*;
+pub use util::events::*;
